Import source code from Debian package
Flyby is to be based on predict-g1yyh, a fork of predict that comes with
Debian's predict packages. This is an import of the complete source code
after applying Debian's patches. It also includes som extra manual pages
and README/copyright information that is included in the Debian package.

This provides a known starting point for further development.

No names have been changed in this patch. The code builds as predict,
not flyby.